Move HashFileStream to lib.exporters
