+++ server/lib/exporters/bag.py
@@ -1,52 +1,10 @@
 import json
-from hashlib import sha256, md5
 from girder.constants import AccessType
 from girder.models.folder import Folder
-from girder.utility import ziputil, hash_state
-from . import default_top_readme, default_bagit
+from girder.utility import ziputil
+from . import default_top_readme, default_bagit, HashFileStream
 from ..license import WholeTaleLicense
 from ..manifest import Manifest
-
-
-class HashFileStream:
-    """Generator that computes md5 and sha256 of data returned by it"""
-
-    def __init__(self, gen):
-        """
-        This class is primarily meant to wrap Girder's download function,
-        which returns iterators, hence self.x = x()
-        """
-        try:
-            self.gen = gen()
-        except TypeError:
-            self.gen = gen
-        self.state = {
-            'md5': hash_state.serializeHex(md5()),
-            'sha256': hash_state.serializeHex(sha256()),
-        }
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        nxt = next(self.gen)
-        for alg in self.state.keys():
-            checksum = hash_state.restoreHex(self.state[alg], alg)
-            checksum.update(nxt)
-            self.state[alg] = hash_state.serializeHex(checksum)
-        return nxt
-
-    def __call__(self):
-        """Needs to be callable, see comment in __init__"""
-        return self
-
-    @property
-    def sha256(self):
-        return hash_state.restoreHex(self.state['sha256'], 'sha256').hexdigest()
-
-    @property
-    def md5(self):
-        return hash_state.restoreHex(self.state['md5'], 'md5').hexdigest()
 
 
 def stream(tale, user):
